+++ cli/commands/filter/interactive.go
@@ -10,10 +10,11 @@ import (
 )
 
 type filterOpts struct {
-	Action     string `survey:"action"`
-	Name       string `survey:"name"`
-	Namespace  string
-	Statements string `survey:"statements"`
+	Action        string `survey:"action"`
+	Name          string `survey:"name"`
+	Namespace     string `survey:"namespace"`
+	Expressions   string `survey:"expressions"`
+	RuntimeAssets string `survey:"runtimeAssets"`
 }
 
 func newFilterOpts() *filterOpts {
@@ -55,12 +56,19 @@ func (opts *filterOpts) administerQuestionnaire(editing bool) error {
 			Validate: survey.Required,
 		},
 		{
-			Name: "statements",
+			Name: "expressions",
 			Prompt: &survey.Input{
-				Message: "Statements (comma separated list):",
-				Default: opts.Statements,
+				Message: "Expressions (comma separated list):",
+				Default: opts.Expressions,
 			},
 			Validate: survey.Required,
+		},
+		{
+			Name: "runtimeAssets",
+			Prompt: &survey.Input{
+				Message: "Runtime Assets",
+				Default: "",
+			},
 		},
 	}...)
 
@@ -71,19 +79,21 @@ func (opts *filterOpts) Copy(filter *types.EventFilter) {
 	filter.Action = opts.Action
 	filter.Name = opts.Name
 	filter.Namespace = opts.Namespace
-	filter.Statements = helpers.SafeSplitCSV(opts.Statements)
+	filter.Expressions = helpers.SafeSplitCSV(opts.Expressions)
+	filter.RuntimeAssets = helpers.SafeSplitCSV(opts.RuntimeAssets)
 }
 
 func (opts *filterOpts) withFilter(filter *types.EventFilter) {
 	opts.Name = filter.Name
 	opts.Namespace = filter.Namespace
 	opts.Action = filter.Action
-	opts.Statements = strings.Join(filter.Statements, ",")
+	opts.Expressions = strings.Join(filter.Expressions, ",")
+	opts.RuntimeAssets = strings.Join(filter.RuntimeAssets, ",")
 }
 
 func (opts *filterOpts) withFlags(flags *pflag.FlagSet) {
 	opts.Action, _ = flags.GetString("action")
-	opts.Statements, _ = flags.GetString("statements")
+	opts.Expressions, _ = flags.GetString("expressions")
 
 	if namespace := helpers.GetChangedStringValueFlag("namespace", flags); namespace != "" {
 		opts.Namespace = namespace