+++ types/event_test.go
@@ -328,27 +328,3 @@ func TestEventsByTimestamp(t *testing.T) {
 		})
 	}
 }
-
-func TestEventGet(t *testing.T) {
-	event := FixtureEvent("entity", "check")
-
-	r, err := event.Get("Timestamp")
-	assert.NoError(t, err)
-	assert.EqualValues(t, event.Timestamp, r)
-
-	r, err = event.Get("Entity")
-	assert.NoError(t, err)
-	assert.EqualValues(t, event.Entity, r)
-
-	r, err = event.Get("Check")
-	assert.NoError(t, err)
-	assert.EqualValues(t, event.Check, r)
-
-	r, err = event.Get("Metrics")
-	assert.NoError(t, err)
-	assert.EqualValues(t, event.Metrics, r)
-
-	r, err = event.Get("Non Existence")
-	assert.Error(t, err)
-	assert.Empty(t, event.Hooks, r)
-}